Move loca_node_params to utils module
Minor keyboard teleoperation node ajustments
